fix(wfs): use bigger integer data type for indices array, for THREE.LineSegments. Prior to this change, there was a bug, visible in the wfs globe example : if you we set buses line WFS layer Level to 10 or less, you will see bugged lines. The bug comes from the big size of the created LineSegments indices array.
This change use a 32 bit ushort instead of a 16 bit ushort for indices array.

+++ src/Renderer/ThreeExtended/Feature2Mesh.js
@@ -176,7 +176,7 @@ function coordinateToLines(coordinates, properties, options) {
 
     geometry.addAttribute('position', new THREE.BufferAttribute(vertices, 3));
     geometry.addAttribute('color', new THREE.BufferAttribute(colors, 3, true));
-    geometry.setIndex(new THREE.BufferAttribute(new Uint16Array(indices), 1));
+    geometry.setIndex(new THREE.BufferAttribute(new Uint32Array(indices), 1));
     return new THREE.LineSegments(geometry);
 }
